+++ Package.swift
@@ -21,10 +21,12 @@ let package = Package(
     dependencies: [
         .package(url: "https://github.com/danielsaidi/BadgeIcon.git", .upToNextMinor(from: "0.6.0")),
         .package(url: "https://github.com/simibac/ConfettiSwiftUI.git", .upToNextMajor(from: "1.1.0")),
+        .package(url: "https://github.com/danielsaidi/FlipKit.git", .upToNextMinor(from: "0.1.0")),
         .package(url: "https://github.com/danielsaidi/OnboardingKit.git", .upToNextMajor(from: "8.3.0")),
-        .package(url: "https://github.com/danielsaidi/StandardButtons.git", .upToNextMajor(from: "0.3.0")),
+        .package(url: "https://github.com/danielsaidi/ScanCodes.git", .upToNextMajor(from: "0.1.0")),
+        .package(url: "https://github.com/danielsaidi/StandardButtons.git", .upToNextMajor(from: "0.7.0")),
         .package(url: "https://github.com/danielsaidi/StoreKitPlus.git", .upToNextMinor(from: "0.7.0")),
-        .package(url: "https://github.com/danielsaidi/SwiftUIKit.git", .upToNextMajor(from: "5.1.1")),
+        .package(url: "https://github.com/danielsaidi/SwiftUIKit.git", .upToNextMajor(from: "5.8.3")),
     ],
     targets: [
         .target(
@@ -32,7 +34,9 @@ let package = Package(
             dependencies: [
                 "BadgeIcon",
                 "ConfettiSwiftUI",
+                "FlipKit",
                 "OnboardingKit",
+                "ScanCodes",
                 "StandardButtons",
                 "StoreKitPlus",
                 "SwiftUIKit"